Extract CocinaGenerator model
The services directory was becoming a confusing mess of code, this helps organize a bit
