+++ .rubocop_todo.yml
@@ -1,23 +1,16 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config --auto-gen-only-exclude`
-# on 2020-11-04 21:37:55 UTC using RuboCop version 0.93.1.
+# on 2021-07-15 16:08:15 UTC using RuboCop version 1.18.3.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
-# Offense count: 2
-# Configuration parameters: IgnoredMethods, Max.
+# Offense count: 4
+# Configuration parameters: IgnoredMethods, CountRepeatedAttributes, Max.
 Metrics/AbcSize:
   Exclude:
     - 'app/controllers/works_controller.rb'
-    - 'app/services/storage_migrator.rb'
-
-# Offense count: 1
-# Configuration parameters: CountComments, Max, CountAsOne, ExcludedMethods.
-Metrics/MethodLength:
-  Exclude:
-    - 'app/services/request_generator.rb'
 
 # Offense count: 1
 RSpec/AnyInstance:
