+++ spec/services/cocina_generator/description/types_generator_spec.rb
@@ -3,7 +3,7 @@
 
 require 'rails_helper'
 
-RSpec.describe TypesGenerator do
+RSpec.describe CocinaGenerator::Description::TypesGenerator do
   let(:work_version) { build(:work_version) }
 
   describe '.generate' do
